refactor pages into components and centralize api
