+++ src/pages/Home.tsx
@@ -1,0 +1,25 @@
+import { Navigation } from '@/components/home/Navigation';
+import { HeroSection } from '@/components/home/HeroSection';
+import { FeaturesSection } from '@/components/home/FeaturesSection';
+import { HowItWorksSection } from '@/components/home/HowItWorksSection';
+import { PricingSection } from '@/components/home/PricingSection';
+import { Footer } from '@/components/home/Footer';
+
+interface HomePageProps {
+  onGetStarted?: () => void;
+}
+
+export function HomePage({ onGetStarted }: HomePageProps) {
+  return (
+    <div className="min-h-screen bg-gradient-to-br from-purple-50 to-purple-100">
+      <Navigation onGetStarted={onGetStarted} />
+      <HeroSection onGetStarted={onGetStarted} />
+      <FeaturesSection />
+      <HowItWorksSection />
+      <PricingSection onGetStarted={onGetStarted} />
+      <Footer />
+    </div>
+  );
+}
+
+export default HomePage;
